Akses Otoritas -> Bagian Kerja Clear

+++ src/app/components/master/bagian-kerja/bagian-kerja.component.ts
@@ -14,6 +14,8 @@ moment.locale('id');
   styleUrls: ['./bagian-kerja.component.css'],
 })
 export class BagianKerjaComponent implements OnInit {
+  akses = this.api.akses.role_bagian_kerja;
+
   table = 'ms_bagiankerja/';
   data!: any;
 
@@ -28,31 +30,33 @@ export class BagianKerjaComponent implements OnInit {
   catchResult: any;
   lokasiValue = '';
 
-  constructor(private api: ApiService, public dialog: MatDialog) {}
+  constructor(private api: ApiService, private dialog: MatDialog) {}
 
   ngOnInit(): void {
     this.getAllData();
   }
 
   tambahData() {
-    this.dialog
-      .open(ModalBagianKerjaComponent)
-      .afterClosed()
-      .subscribe((result) => {
-        if (result !== undefined) {
-          this.api.postData(this.table, result).subscribe(() => {
-            this.length = this.length + 1;
-            this.getPageData();
-          });
-        }
-      });
+    if (this.akses.edit) {
+      this.dialog
+        .open(ModalBagianKerjaComponent)
+        .afterClosed()
+        .subscribe((result) => {
+          if (result !== undefined) {
+            this.api.postData(this.table, result).subscribe(() => {
+              this.length = this.length + 1;
+              this.getPageData();
+            });
+          }
+        });
+    } else {
+      window.alert('Anda tidak memiliki Akses');
+    }
   }
 
   editData(data: any) {
     this.dialog
-      .open(ModalBagianKerjaComponent, {
-        data: data,
-      })
+      .open(ModalBagianKerjaComponent, { data: data })
       .afterClosed()
       .subscribe((result) => {
         if (result !== undefined)
@@ -63,17 +67,21 @@ export class BagianKerjaComponent implements OnInit {
   }
 
   deleteData(id: number) {
-    this.dialog
-      .open(VoidComponent)
-      .afterClosed()
-      .subscribe((result) => {
-        if (result === 'ya') {
-          this.api.deleteData(this.table + id).subscribe(() => {
-            this.length = this.length - 1;
-            this.getPageData();
-          });
-        }
-      });
+    if (this.akses.edit) {
+      this.dialog
+        .open(VoidComponent)
+        .afterClosed()
+        .subscribe((result) => {
+          if (result === 'ya') {
+            this.api.deleteData(this.table + id).subscribe(() => {
+              this.length = this.length - 1;
+              this.getPageData();
+            });
+          }
+        });
+    } else {
+      window.alert('Anda tidak memiliki Akses');
+    }
   }
 
   handlePageEvent(event: PageEvent) {
@@ -150,41 +158,45 @@ export class BagianKerjaComponent implements OnInit {
   }
 
   printData() {
-    let content = this.data.map((res: any) => ({
-      Jenis: res.jenis_bagian,
-      Lokasi: res.lokasi,
-      Uplink: res.divisi === undefined ? res.departemen : res.divisi,
-      Deskripsi: res.sub_departemen,
-    }));
-    const ws = utils.json_to_sheet(content);
-    const wsTemp = utils.json_to_sheet(content);
-    let length = Number(ws['!ref']?.slice(4));
-    let cell = 4;
+    if (this.akses.download) {
+      let content = this.data.map((res: any) => ({
+        Jenis: res.jenis_bagian,
+        Lokasi: res.lokasi,
+        Uplink: res.divisi === undefined ? res.departemen : res.divisi,
+        Deskripsi: res.sub_departemen,
+      }));
+      const ws = utils.json_to_sheet(content);
+      const wsTemp = utils.json_to_sheet(content);
+      let length = Number(ws['!ref']?.slice(4));
+      let cell = 4;
 
-    ws['!ref'] = 'A1:D' + (length + cell);
-    for (let i = 1; i <= 4; i++) {
-      ws['A' + i] = { t: 's', v: '' };
-      ws['B' + i] = { t: 's', v: '' };
-      ws['C' + i] = { t: 's', v: '' };
-      ws['D' + i] = { t: 's', v: '' };
+      ws['!ref'] = 'A1:D' + (length + cell);
+      for (let i = 1; i <= 4; i++) {
+        ws['A' + i] = { t: 's', v: '' };
+        ws['B' + i] = { t: 's', v: '' };
+        ws['C' + i] = { t: 's', v: '' };
+        ws['D' + i] = { t: 's', v: '' };
+      }
+
+      ws['A1'] = { t: 's', v: 'Bagian Kerja' };
+      ws['C1'] = { t: 's', v: 'Tanggal Cetak' };
+      ws['C2'] = { t: 's', v: 'User :' };
+      ws['D1'] = { t: 's', v: moment().format('DD MMM YYYY') };
+      ws['D2'] = { t: 's', v: window.localStorage.getItem('key') };
+
+      for (let i = 0; i < length; i++) {
+        ws['A' + (i + cell)] = wsTemp['A' + (i + 1)];
+        ws['B' + (i + cell)] = wsTemp['B' + (i + 1)];
+        ws['C' + (i + cell)] = wsTemp['C' + (i + 1)];
+        ws['D' + (i + cell)] = wsTemp['D' + (i + 1)];
+      }
+
+      const wb = utils.book_new();
+
+      utils.book_append_sheet(wb, ws);
+      writeFileXLSX(wb, 'Bagian Kerja.xlsx');
+    } else {
+      window.alert('Anda tidak memiliki Akses');
     }
-
-    ws['A1'] = { t: 's', v: 'Bagian Kerja' };
-    ws['C1'] = { t: 's', v: 'Tanggal Cetak' };
-    ws['C2'] = { t: 's', v: 'User :' };
-    ws['D1'] = { t: 's', v: moment().format('DD MMM YYYY') };
-    ws['D2'] = { t: 's', v: window.localStorage.getItem('key') };
-
-    for (let i = 0; i < length; i++) {
-      ws['A' + (i + cell)] = wsTemp['A' + (i + 1)];
-      ws['B' + (i + cell)] = wsTemp['B' + (i + 1)];
-      ws['C' + (i + cell)] = wsTemp['C' + (i + 1)];
-      ws['D' + (i + cell)] = wsTemp['D' + (i + 1)];
-    }
-
-    const wb = utils.book_new();
-
-    utils.book_append_sheet(wb, ws);
-    writeFileXLSX(wb, 'Bagian Kerja.xlsx');
   }
 }
